Revert "fix NNS distance for new cell types"
