+++ searchlib/src/vespa/searchlib/tensor/distance_function_factory.cpp
@@ -2,10 +2,6 @@
 
 #include "distance_function_factory.h"
 #include "distance_functions.h"
-#include <vespa/vespalib/util/typify.h>
-#include <vespa/log/log.h>
-
-LOG_SETUP(".searchlib.tensor.distance_function_factory");
 
 using search::attribute::DistanceMetric;
 using vespalib::eval::CellType;
@@ -17,28 +13,41 @@ DistanceFunction::UP
 make_distance_function(DistanceMetric variant, CellType cell_type)
 {
     switch (variant) {
-    case DistanceMetric::Euclidean:
-        switch (cell_type) {
-        case CellType::FLOAT:  return std::make_unique<SquaredEuclideanDistanceHW<float>>();
-        case CellType::DOUBLE: return std::make_unique<SquaredEuclideanDistanceHW<double>>();
-        default:               return std::make_unique<SquaredEuclideanDistance>();
-        } 
-    case DistanceMetric::Angular:
-        switch (cell_type) {
-        case CellType::FLOAT:  return std::make_unique<AngularDistanceHW<float>>();
-        case CellType::DOUBLE: return std::make_unique<AngularDistanceHW<double>>();
-        default:               return std::make_unique<AngularDistance>();
-        }
-    case DistanceMetric::GeoDegrees:
-        return std::make_unique<GeoDegreesDistance>();
-    case DistanceMetric::InnerProduct:
-        switch (cell_type) {
-        case CellType::FLOAT:  return std::make_unique<InnerProductDistanceHW<float>>();
-        case CellType::DOUBLE: return std::make_unique<InnerProductDistanceHW<double>>();
-        default:               return std::make_unique<InnerProductDistance>();
-        }
-    case DistanceMetric::Hamming:
-        return std::make_unique<HammingDistance>();
+        case DistanceMetric::Euclidean:
+            if (cell_type == CellType::FLOAT) {
+                return std::make_unique<SquaredEuclideanDistance<float>>();
+            } else {
+                return std::make_unique<SquaredEuclideanDistance<double>>();
+            }
+            break;
+        case DistanceMetric::Angular:
+            if (cell_type == CellType::FLOAT) {
+                return std::make_unique<AngularDistance<float>>();
+            } else {
+                return std::make_unique<AngularDistance<double>>();
+            }
+            break;
+        case DistanceMetric::GeoDegrees:
+            if (cell_type == CellType::FLOAT) {
+                return std::make_unique<GeoDegreesDistance<float>>();
+            } else {
+                return std::make_unique<GeoDegreesDistance<double>>();
+            }
+            break;
+        case DistanceMetric::InnerProduct:
+            if (cell_type == CellType::FLOAT) {
+                return std::make_unique<InnerProductDistance<float>>();
+            } else {
+                return std::make_unique<InnerProductDistance<double>>();
+            }
+            break;
+        case DistanceMetric::Hamming:
+            if (cell_type == CellType::FLOAT) {
+                return std::make_unique<HammingDistance<float>>();
+            } else {
+                return std::make_unique<HammingDistance<double>>();
+            }
+            break;
     }
     // not reached:
     return DistanceFunction::UP();
